requirements.txt
yfinance>=0.2.36
openai>=1.12.0
python-telegram-bot>=21.0
supabase>=2.3.0
tenacity>=8.2.0
structlog>=24.1.0
feedparser>=6.0.0
pandas>=2.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
